feat(mesh): ADR-049 Phase 5 — move generic security primitives into hive-mesh
Move protocol-agnostic security modules from hive-protocol into
hive-mesh so the mesh layer has zero dependency on hive-protocol
for core crypto operations.

New hive-mesh/src/security/ modules:
  - error.rs      — SecurityError enum
  - device_id.rs  — DeviceId (Ed25519 fingerprint → NodeId bridge)
  - keypair.rs    — DeviceKeypair (Ed25519 signing)
  - callsign.rs   — CallsignGenerator (NATO phonetic alphabet)
  - encryption.rs — EncryptionKeypair, EncryptionManager, SecureChannel, GroupKey
  - formation_key.rs — FormationKey (HMAC-SHA256 pre-shared key)

hive-protocol/src/security/ files become thin re-export stubs that
forward to hive_mesh::security, preserving full API compatibility.

Adds crypto deps to hive-mesh: x25519-dalek, hkdf, hmac, sha2,
subtle, base64, hex, rand_core.
