+++ hive-mesh/Cargo.toml
@@ -26,12 +26,21 @@ uuid = { version = "1.0", features = ["v4", "serde"] }
 ed25519-dalek = { version = "2", features = ["rand_core"] }
 chacha20poly1305 = "0.10"
 
+# Security primitives (ADR-049 Phase 5)
+x25519-dalek = { version = "2", features = ["static_secrets"] }
+hkdf = "0.12"
+hmac = "0.12"
+sha2 = "0.10"
+subtle = "2.6"
+base64 = "0.22"
+hex = "0.4"
+rand_core = { version = "0.6", features = ["getrandom"] }
+
 # Discovery (ADR-049 Phase 2)
 mdns-sd = "0.11"
 toml = "0.8"
 
 # HIVE-Lite transport (ADR-035, optional)
-hex = { version = "0.4", optional = true }
 log = { version = "0.4", optional = true }
 
 # Bluetooth LE transport (ADR-032/ADR-039, optional)
@@ -44,14 +53,13 @@ iroh-blobs = { version = "0.97", optional = true }
 redb = { version = "2.4", optional = true }
 lru = { version = "0.12", optional = true }
 negentropy = { version = "0.5", optional = true }
-sha2 = { version = "0.10", optional = true }
 rand = { version = "0.9", optional = true }
 
 [features]
 default = []
-lite-transport = ["hex", "log"]
+lite-transport = ["log"]
 bluetooth = ["hive-btle"]
-automerge-backend = ["automerge", "iroh", "iroh-blobs", "redb", "lru", "negentropy", "hex", "sha2", "rand"]
+automerge-backend = ["automerge", "iroh", "iroh-blobs", "redb", "lru", "negentropy", "rand"]
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["test-util", "macros"] }
